Shift TypicalPerson.java into TestData.java for easier readability
